feat: Refactor routes that are not split with methods
